refactor: 💡 improve implement of table dnd
✅ Closes: #2014
