+++ packages/plugins/preset-gfm/src/node/table/utils/index.ts
@@ -1,0 +1,11 @@
+export * from './create-table'
+export * from './find-table'
+export * from './get-cells-in-col'
+export * from './get-cells-in-row'
+export * from './select-line'
+export * from './types'
+export * from './add-row-with-alignment'
+export * from './move-row'
+export * from './move-col'
+export * from './select-table'
+export * from './get-all-cells-in-table'
